backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
